Initial commit.
This is a copy (essentially) of the github repo (as of 1 April 2015)
and is intended to be used as a basis for the externally visible sdk
and as a parent for the internal sdk.

We should be able to merge changes to this in to the github repo as
products and apis become available.