link update
Updated broken links and did minor formatting. Attempted to add *live* to telesign.api.PhoneId - tool continually displays added phoneid live info with different formatting and as deleted (crossed out) at the bottom of the table when added. When tested, this happened on the bottom of each table section. 

Hoping that this is a preview bug and it will post with added content, but not sure how it will come out.